The picker stays in the same place when the hue slider is moved
When the hue slider is moved, two colors are now created. One fully saturated and bright value for the picker area, and one color where only the Hue from the HSV triplet has been changed. The latter is used for the callbacks.

Additionally, the picker dot is no longer moved to the top right corner when the hue slider is moved.